requirements.txt
Flask==3.0.3
flask_cors==4.0.1
Pillow==10.4.0
Requests==2.32.3
